+++ tests/rule_based_profiler/conftest.py
@@ -6,16 +6,16 @@ from typing import Any, Dict, List
 import numpy as np
 import pandas as pd
 import pytest
+from freezegun import freeze_time
 from ruamel.yaml import YAML
 
 from great_expectations import DataContext
+from great_expectations.core import ExpectationConfiguration, ExpectationSuite
 from great_expectations.data_context.util import file_relative_path
 from great_expectations.datasource.data_connector.util import (
     get_filesystem_one_level_directory_glob_path_list,
 )
 from great_expectations.execution_engine.execution_engine import MetricDomainTypes
-
-# noinspection PyUnresolvedReferences
 from great_expectations.rule_based_profiler.domain_builder import Domain
 from great_expectations.rule_based_profiler.parameter_builder import (
     ParameterContainer,
@@ -23,22 +23,209 @@ from great_expectations.rule_based_profiler.parameter_builder import (
 )
 from great_expectations.rule_based_profiler.rule import Rule
 
-# noinspection PyUnresolvedReferences
-from tests.rule_based_profiler.alice_user_workflow_fixture import (
-    alice_columnar_table_single_batch,
-)
-
-# noinspection PyUnresolvedReferences
-from tests.rule_based_profiler.bobby_user_workflow_fixture import (
-    bobby_columnar_table_multi_batch,
-)
-
-# noinspection PyUnresolvedReferences
-from tests.rule_based_profiler.bobster_user_workflow_fixture import (
-    bobster_columnar_table_multi_batch_normal_mean_5000_stdev_1000,
-)
-
 yaml = YAML()
+
+
+@pytest.fixture
+@freeze_time("09/26/2019 13:42:41")
+def alice_columnar_table_single_batch(empty_data_context):
+    """
+    About the "Alice" User Workflow Fixture
+
+    Alice has a single table of columnar data called user_events (DataAsset) that she wants to check periodically as new
+    data is added.
+
+      - She knows what some of the columns mean, but not all - and there are MANY of them (only a subset currently shown
+        in examples and fixtures).
+
+      - She has organized other tables similarly so that for example column name suffixes indicate which are for user
+        ids (_id) and which timestamps are for versioning (_ts).
+
+    She wants to use a configurable profiler to generate a description (ExpectationSuite) about table so that she can:
+
+        1. use it to validate the user_events table periodically and set up alerts for when things change
+
+        2. have a place to add her domain knowledge of the data (that can also be validated against new data)
+
+        3. if all goes well, generalize some of the Profiler to use on her other tables
+
+    Alice configures her Profiler using the YAML configurations and data file locations captured in this fixture.
+    """
+
+    verbose_profiler_config_file_path: str = file_relative_path(
+        __file__,
+        os.path.join(
+            "..",
+            "test_fixtures",
+            "rule_based_profiler",
+            "alpha",
+            "alice_user_workflow_verbose_profiler_config.yml",
+        ),
+    )
+
+    verbose_profiler_config: str
+    with open(verbose_profiler_config_file_path) as f:
+        verbose_profiler_config = f.read()
+
+    my_rule_for_user_ids_expectation_configurations: List[ExpectationConfiguration] = [
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_values_to_be_of_type",
+                "kwargs": {
+                    "column": "user_id",
+                    "type_": "INTEGER",
+                },
+                "meta": {},
+            }
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_values_to_be_between",
+                "kwargs": {
+                    "min_value": 397433,  # From the data
+                    "max_value": 999999999999,
+                    "column": "user_id",
+                },
+                "meta": {},
+            }
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_values_to_not_be_null",
+                "kwargs": {
+                    "column": "user_id",
+                },
+                "meta": {},
+            }
+        ),
+    ]
+
+    event_ts_column_data: Dict[str, str] = {
+        "column_name": "event_ts",
+        "observed_max_time_str": "2004-10-19 11:05:20",
+    }
+
+    my_rule_for_timestamps_column_data: List[Dict[str, str]] = [
+        event_ts_column_data,
+        {
+            "column_name": "server_ts",
+            "observed_max_time_str": "2004-10-19 11:05:20",
+        },
+        {
+            "column_name": "device_ts",
+            "observed_max_time_str": "2004-10-19 11:05:22",
+        },
+    ]
+    my_rule_for_timestamps_expectation_configurations: List[
+        ExpectationConfiguration
+    ] = []
+    column_data: Dict[str, str]
+    for column_data in my_rule_for_timestamps_column_data:
+        my_rule_for_timestamps_expectation_configurations.extend(
+            [
+                ExpectationConfiguration(
+                    **{
+                        "expectation_type": "expect_column_values_to_be_of_type",
+                        "kwargs": {
+                            "column": column_data["column_name"],
+                            "type_": "TIMESTAMP",
+                        },
+                        "meta": {},
+                    }
+                ),
+                ExpectationConfiguration(
+                    **{
+                        "expectation_type": "expect_column_values_to_be_increasing",
+                        "kwargs": {
+                            "column": column_data["column_name"],
+                        },
+                        "meta": {},
+                    }
+                ),
+                ExpectationConfiguration(
+                    **{
+                        "expectation_type": "expect_column_values_to_be_dateutil_parseable",
+                        "kwargs": {
+                            "column": column_data["column_name"],
+                        },
+                        "meta": {},
+                    }
+                ),
+                ExpectationConfiguration(
+                    **{
+                        "expectation_type": "expect_column_min_to_be_between",
+                        "kwargs": {
+                            "column": column_data["column_name"],
+                            "min_value": "2004-10-19T10:23:54",  # From variables
+                            "max_value": "2004-10-19T10:23:54",  # From variables
+                        },
+                        "meta": {
+                            "notes": {
+                                "format": "markdown",
+                                "content": [
+                                    "### This expectation confirms no events occur before tracking started **2004-10-19 10:23:54**"
+                                ],
+                            }
+                        },
+                    }
+                ),
+                ExpectationConfiguration(
+                    **{
+                        "expectation_type": "expect_column_max_to_be_between",
+                        "kwargs": {
+                            "column": column_data["column_name"],
+                            "min_value": "2004-10-19T10:23:54",  # From variables
+                            "max_value": event_ts_column_data[
+                                "observed_max_time_str"
+                            ],  # Pin to event_ts column
+                        },
+                        "meta": {
+                            "notes": {
+                                "format": "markdown",
+                                "content": [
+                                    "### This expectation confirms that the event_ts contains the latest timestamp of all domains"
+                                ],
+                            }
+                        },
+                    }
+                ),
+            ]
+        )
+
+    expectation_configurations: List[ExpectationConfiguration] = []
+
+    expectation_configurations.extend(my_rule_for_user_ids_expectation_configurations)
+    expectation_configurations.extend(my_rule_for_timestamps_expectation_configurations)
+
+    expectation_suite_name: str = "alice_columnar_table_single_batch"
+    expected_expectation_suite: ExpectationSuite = ExpectationSuite(
+        expectation_suite_name=expectation_suite_name, data_context=empty_data_context
+    )
+    expectation_configuration: ExpectationConfiguration
+    for expectation_configuration in expectation_configurations:
+        # NOTE Will 20211208 add_expectation() method, although being called by an ExpectationSuite instance, is being
+        # called within a fixture, and we will prevent it from sending a usage_event by calling the private method
+        # _add_expectation().
+        expected_expectation_suite._add_expectation(
+            expectation_configuration=expectation_configuration, send_usage_event=False
+        )
+
+    # NOTE that this expectation suite should fail when validated on the data in "sample_data_relative_path"
+    # because the device_ts is ahead of the event_ts for the latest event
+    sample_data_relative_path: str = "alice_columnar_table_single_batch_data.csv"
+
+    profiler_config: dict = yaml.load(verbose_profiler_config)
+    expected_expectation_suite.add_citation(
+        comment="Suite created by Rule-Based Profiler with the configuration included.",
+        profiler_config=profiler_config,
+    )
+
+    return {
+        "profiler_config": verbose_profiler_config,
+        "expected_expectation_suite_name": expectation_suite_name,
+        "expected_expectation_suite": expected_expectation_suite,
+        "sample_data_relative_path": sample_data_relative_path,
+    }
 
 
 @pytest.fixture
@@ -132,6 +319,713 @@ data_connectors:
 
 
 @pytest.fixture
+@freeze_time("09/26/2019 13:42:41")
+def bobby_columnar_table_multi_batch(empty_data_context):
+    """
+    About the "Bobby" User Workflow Fixture
+
+    Bobby has multiple tables of columnar data called user_events (DataAsset) that he wants to check periodically as new
+    data is added.
+
+      - He knows what some of the columns are of the acconting/financial/account type.
+
+    He wants to use a configurable profiler to generate a description (ExpectationSuite) about tables so that he can:
+
+        1. monitor the average number of rows in the tables
+
+        2. use it to validate min/max boundaries of all columns are of the acconting/financial/account type and set up alerts for when things change
+
+        3. have a place to add his domain knowledge of the data (that can also be validated against new data)
+
+        4. if all goes well, generalize some of the Profiler to use on his other tables
+
+    Bobby uses a crude, highly inaccurate deterministic parametric estimator -- for illustrative purposes.
+
+    Bobby configures his Profiler using the YAML configurations and data file locations captured in this fixture.
+    """
+
+    verbose_profiler_config_file_path: str = file_relative_path(
+        __file__,
+        os.path.join(
+            "..",
+            "test_fixtures",
+            "rule_based_profiler",
+            "alpha",
+            "bobby_user_workflow_verbose_profiler_config.yml",
+        ),
+    )
+
+    verbose_profiler_config: str
+    with open(verbose_profiler_config_file_path) as f:
+        verbose_profiler_config = f.read()
+
+    my_row_count_range_rule_expectation_configurations_oneshot_sampling_method: List[
+        ExpectationConfiguration
+    ] = [
+        ExpectationConfiguration(
+            **{
+                "kwargs": {"min_value": 7505, "max_value": 8495},
+                "expectation_type": "expect_table_row_count_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "table.row_count",
+                            "domain_kwargs": {},
+                        },
+                        "num_batches": 2,
+                    },
+                },
+            },
+        ),
+    ]
+
+    my_column_ranges_rule_expectation_configurations_oneshot_sampling_method: List[
+        ExpectationConfiguration
+    ] = [
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "VendorID"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "VendorID",
+                    "min_value": 1,
+                    "max_value": 1,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "VendorID"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "VendorID",
+                    "min_value": 4,
+                    "max_value": 4,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "passenger_count"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "passenger_count",
+                    "min_value": 0,
+                    "max_value": 1,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "passenger_count"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "passenger_count",
+                    "min_value": 6,
+                    "max_value": 6,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "trip_distance"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "trip_distance",
+                    "min_value": 0.0,
+                    "max_value": 0.0,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "trip_distance"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "trip_distance",
+                    "min_value": 37.62,
+                    "max_value": 57.85,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "RatecodeID"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "RatecodeID",
+                    "min_value": 1,
+                    "max_value": 1,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "RatecodeID"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "RatecodeID",
+                    "min_value": 5,
+                    "max_value": 6,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "PULocationID"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "PULocationID",
+                    "min_value": 1,
+                    "max_value": 1,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "PULocationID"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "PULocationID",
+                    "min_value": 265,
+                    "max_value": 265,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "DOLocationID"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "DOLocationID",
+                    "min_value": 1,
+                    "max_value": 1,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "DOLocationID"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "DOLocationID",
+                    "min_value": 265,
+                    "max_value": 265,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "payment_type"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "payment_type",
+                    "min_value": 1,
+                    "max_value": 1,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "payment_type"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "payment_type",
+                    "min_value": 4,
+                    "max_value": 4,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "fare_amount"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "fare_amount",
+                    "min_value": -51.84,
+                    "max_value": -21.16,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "fare_amount"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "fare_amount",
+                    "min_value": 228.94,
+                    "max_value": 2990.05,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "extra"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "extra",
+                    "min_value": -36.53,
+                    "max_value": -1.18,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "extra"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "extra",
+                    "min_value": 4.51,
+                    "max_value": 6.99,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "mta_tax"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "mta_tax",
+                    "min_value": -0.5,
+                    "max_value": -0.5,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "mta_tax"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "mta_tax",
+                    "min_value": 0.69,
+                    "max_value": 37.32,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "tip_amount"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "tip_amount",
+                    "min_value": 0.0,
+                    "max_value": 0.0,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "tip_amount"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "tip_amount",
+                    "min_value": 46.84,
+                    "max_value": 74.86,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "tolls_amount"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "tolls_amount",
+                    "min_value": 0.0,
+                    "max_value": 0.0,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "tolls_amount"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "tolls_amount",
+                    "min_value": 26.4,
+                    "max_value": 497.67,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "improvement_surcharge"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "improvement_surcharge",
+                    "min_value": -0.3,
+                    "max_value": -0.3,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "improvement_surcharge"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "improvement_surcharge",
+                    "min_value": 0.3,
+                    "max_value": 0.3,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "total_amount"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "total_amount",
+                    "min_value": -52.66,
+                    "max_value": -24.44,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "total_amount"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "total_amount",
+                    "min_value": 550.18,
+                    "max_value": 2992.47,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_min_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.min",
+                            "domain_kwargs": {"column": "congestion_surcharge"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "congestion_surcharge",
+                    "min_value": -2.49,
+                    "max_value": -0.01,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+        ExpectationConfiguration(
+            **{
+                "expectation_type": "expect_column_max_to_be_between",
+                "meta": {
+                    "profiler_details": {
+                        "metric_configuration": {
+                            "metric_name": "column.max",
+                            "domain_kwargs": {"column": "congestion_surcharge"},
+                        },
+                        "num_batches": 2,
+                    }
+                },
+                "kwargs": {
+                    "column": "congestion_surcharge",
+                    "min_value": 0.01,
+                    "max_value": 2.49,
+                    "mostly": 1.0,
+                },
+            },
+        ),
+    ]
+
+    expectation_configurations: List[ExpectationConfiguration] = []
+
+    expectation_configurations.extend(
+        my_row_count_range_rule_expectation_configurations_oneshot_sampling_method
+    )
+    expectation_configurations.extend(
+        my_column_ranges_rule_expectation_configurations_oneshot_sampling_method
+    )
+
+    expectation_suite_name_oneshot_sampling_method: str = (
+        "bobby_columnar_table_multi_batch_oneshot_sampling_method"
+    )
+    expected_expectation_suite_oneshot_sampling_method: ExpectationSuite = (
+        ExpectationSuite(
+            expectation_suite_name=expectation_suite_name_oneshot_sampling_method,
+            data_context=empty_data_context,
+        )
+    )
+    expectation_configuration: ExpectationConfiguration
+    for expectation_configuration in expectation_configurations:
+        # NOTE Will 20211208 add_expectation() method, although being called by an ExpectationSuite instance, is being
+        # called within a fixture, and we will prevent it from sending a usage_event by calling the private method.
+        expected_expectation_suite_oneshot_sampling_method._add_expectation(
+            expectation_configuration=expectation_configuration, send_usage_event=False
+        )
+
+    profiler_config: dict = yaml.load(verbose_profiler_config)
+    expected_expectation_suite_oneshot_sampling_method.add_citation(
+        comment="Suite created by Rule-Based Profiler with the configuration included.",
+        profiler_config=profiler_config,
+    )
+
+    return {
+        "profiler_config": verbose_profiler_config,
+        "test_configuration_oneshot_sampling_method": {
+            "expectation_suite_name": expectation_suite_name_oneshot_sampling_method,
+            "expected_expectation_suite": expected_expectation_suite_oneshot_sampling_method,
+        },
+    }
+
+
+@pytest.fixture
 def bobby_columnar_table_multi_batch_deterministic_data_context(
     tmp_path_factory,
     monkeypatch,
@@ -217,6 +1111,90 @@ def bobby_columnar_table_multi_batch_deterministic_data_context(
 
 
 @pytest.fixture
+def bobster_columnar_table_multi_batch_normal_mean_5000_stdev_1000():
+    """
+    About the "Bobster" User Workflow Fixture
+
+    Bobster has multiple tables of columnar data called user_events (DataAsset) that he wants to check periodically as
+    new data is added.
+
+      - He knows what some of the columns are of the acconting/financial/account type, but he is currently interested in
+        the average table size (in terms of the number of rows in a table).
+
+    He wants to use a configurable profiler to generate a description (ExpectationSuite) about tables so that he can:
+
+        1. monitor the average number of rows in the tables
+
+        2. have a place to add his domain knowledge of the data (that can also be validated against new data)
+
+        3. if all goes well, generalize some of the Profiler to use on his other tables
+
+    Bobster uses a custom implementation of the "bootstrap" non-parametric (i.e, data-driven) statistical estimator.
+
+    Bobster configures his Profiler using the YAML configurations and data file locations captured in this fixture.
+    """
+
+    verbose_profiler_config_file_path: str = file_relative_path(
+        __file__,
+        os.path.join(
+            "..",
+            "test_fixtures",
+            "rule_based_profiler",
+            "alpha",
+            "bobster_user_workflow_verbose_profiler_config.yml",
+        ),
+    )
+
+    verbose_profiler_config: str
+    with open(verbose_profiler_config_file_path) as f:
+        verbose_profiler_config = f.read()
+
+    expectation_suite_name_bootstrap_sampling_method: str = (
+        "bobby_columnar_table_multi_batch_bootstrap_sampling_method"
+    )
+
+    my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_mean_value: int = (
+        5000
+    )
+    my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_std_value: float = (
+        1.0e3
+    )
+    my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_num_stds: float = (
+        3.00
+    )
+
+    my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_min_value_mean_value: int = round(
+        float(
+            my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_mean_value
+        )
+        - (
+            my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_num_stds
+            * my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_std_value
+        )
+    )
+
+    my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_max_value_mean_value: int = round(
+        float(
+            my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_mean_value
+        )
+        + (
+            my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_num_stds
+            * my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_std_value
+        )
+    )
+
+    return {
+        "profiler_config": verbose_profiler_config,
+        "test_configuration_bootstrap_sampling_method": {
+            "expectation_suite_name": expectation_suite_name_bootstrap_sampling_method,
+            "expect_table_row_count_to_be_between_mean_value": my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_mean_value,
+            "expect_table_row_count_to_be_between_min_value_mean_value": my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_min_value_mean_value,
+            "expect_table_row_count_to_be_between_max_value_mean_value": my_row_count_range_rule_expect_table_row_count_to_be_between_expectation_max_value_mean_value,
+        },
+    }
+
+
+@pytest.fixture
 def bobster_columnar_table_multi_batch_normal_mean_5000_stdev_1000_data_context(
     tmp_path_factory,
     monkeypatch,
@@ -283,6 +1261,7 @@ def bobster_columnar_table_multi_batch_normal_mean_5000_stdev_1000_data_context(
         df = df.sample(
             n=output_file_name_length_map[file_name], replace=False, random_state=1
         )
+        # noinspection PyTypeChecker
         df.to_csv(
             path_or_buf=os.path.join(context_path, "..", "data", file_name), index=False
         )
@@ -521,7 +1500,6 @@ def parameter_values_eight_parameters_multiple_depths():
     return parameter_values
 
 
-# noinspection PyPep8Naming
 @pytest.fixture
 def rule_without_variables_without_parameters():
     rule: Rule = Rule(
@@ -534,7 +1512,6 @@ def rule_without_variables_without_parameters():
     return rule
 
 
-# noinspection PyPep8Naming
 @pytest.fixture
 def rule_with_variables_with_parameters(
     column_Age_domain,
